+++ src/main/java/com/moandjiezana/toml/ArrayConverter.java
@@ -1,18 +1,16 @@
 package com.moandjiezana.toml;
 
 import static com.moandjiezana.toml.ValueConverterUtils.INVALID;
-import static com.moandjiezana.toml.ValueConverterUtils.parse;
-import static com.moandjiezana.toml.ValueConverterUtils.parser;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
 class ArrayConverter implements ValueConverter {
 
   static final ArrayConverter ARRAY_PARSER = new ArrayConverter();
 
-  private static final List<Object> INVALID_ARRAY = new ArrayList<Object>();
-  private static final ValueConverters VALUE_ANALYSIS = new ValueConverters();
+  private static final ValueConverters VALUE_CONVERTERS = new ValueConverters();
 
   @Override
   public boolean canConvert(String s) {
@@ -21,37 +19,105 @@ class ArrayConverter implements ValueConverter {
 
   @Override
   public Object convert(String s) {
-    List<Object> tokens = parse(parser().Array(), s);
-    List<Object> values = convertList(tokens);
+    return convert(s, new AtomicInteger(1), true);
+  }
+  
+  public Object convert(String s, AtomicInteger sharedIndex, boolean topLevel) {
+    char[] chars = s.toCharArray();
+    List<Object> arrayItems = new ArrayList<Object>();
+    boolean terminated = false;
+    StringType stringType = StringType.NONE;
+    StringBuilder current = new StringBuilder();
+    
+    for (int i = 1; i < chars.length; i++, sharedIndex.incrementAndGet()) {
+      char c = chars[i];
 
-    if (values == INVALID_ARRAY) {
+      if (terminated && !topLevel) {
+        break;
+      }
+
+      if (terminated) {
+        if (c == '#') {
+          break;
+        }
+        if (!Character.isWhitespace(c)) {
+          return INVALID;
+        }
+        continue;
+      }
+
+      if (stringType == StringType.NONE) {
+        if (c == ',') {
+          if (current.toString().trim().length() > 0) {
+            arrayItems.add(current.toString());
+          }
+          current = new StringBuilder();
+          continue;
+        }
+
+        if (c == '[') {
+          arrayItems.add(convert(s.substring(i), sharedIndex, false));
+          i = sharedIndex.get();
+          continue;
+        }
+
+        if (c == ']') {
+          terminated = true;
+          if (current.toString().trim().length() > 0) {
+            arrayItems.add(current.toString());
+          }
+          current = new StringBuilder();
+          continue;
+        }
+      }
+
+      if (c == '"' && chars[i - 1] != '\\' && !stringType.accepts(c)) {
+        if (chars.length > i + 2 && chars[i + 1] == c && chars[i + 2] == c) {
+          stringType = stringType.flip(StringType.MULTILINE);
+        } else {
+          stringType = stringType.flip(StringType.BASIC);
+        }
+      }
+      
+      if (c == '\'' && !stringType.accepts(c)) {
+        if (chars.length > i + 2 && chars[i + 1] == c && chars[i + 2] == c) {
+          stringType = stringType.flip(StringType.MULTILINE_LITERAL);
+        } else {
+          stringType = stringType.flip(StringType.LITERAL);
+        }
+      }
+
+      current.append(c);
+    }
+    
+    if (!terminated) {
       return INVALID;
     }
-
-    return values;
+    
+    return convertList(arrayItems);
   }
 
-  private List<Object> convertList(List<Object> tokens) {
+  private Object convertList(List<Object> tokens) {
     ArrayList<Object> nestedList = new ArrayList<Object>();
 
     for (Object token : tokens) {
       if (token instanceof String) {
-        Object converted = VALUE_ANALYSIS.convert(((String) token).trim());
+        Object converted = VALUE_CONVERTERS.convert(((String) token).trim());
         if (converted == INVALID) {
-          return INVALID_ARRAY;
+          return INVALID;
         }
         if (isHomogenousArray(converted, nestedList)) {
           nestedList.add(converted);
         } else {
-          return INVALID_ARRAY;
+          return INVALID;
         }
       } else if (token instanceof List) {
         @SuppressWarnings("unchecked")
-        List<Object> convertedList = convertList((List<Object>) token);
-        if (convertedList != INVALID_ARRAY && isHomogenousArray(convertedList, nestedList)) {
+        List<Object> convertedList = (List<Object>) token;
+        if (isHomogenousArray(convertedList, nestedList)) {
           nestedList.add(convertedList);
         } else {
-          return INVALID_ARRAY;
+          return INVALID;
         }
       }
     }
@@ -61,6 +127,26 @@ class ArrayConverter implements ValueConverter {
 
   private boolean isHomogenousArray(Object o, List<?> values) {
     return values.isEmpty() || values.get(0).getClass().isAssignableFrom(o.getClass()) || o.getClass().isAssignableFrom(values.get(0).getClass());
+  }
+  
+  private static enum StringType {
+    NONE, BASIC, LITERAL, MULTILINE, MULTILINE_LITERAL;
+    
+    StringType flip(StringType to) {
+      return this == NONE ? to : NONE;
+    }
+    
+    boolean accepts(char c) {
+      if (this == BASIC || this == MULTILINE) {
+        return c != '"';
+      }
+      
+      if (this == LITERAL || this == MULTILINE_LITERAL) {
+        return c != '\'';
+      }
+      
+      return false;
+    }
   }
 
   private ArrayConverter() {}